feat: add native bounded macOS recursive search
