--- .github/workflows/apply-macos-recursive-search.yml
@@ -1,37 +1,0 @@
-name: Apply macOS Recursive Search
-
-on:
-  push:
-    branches:
-      - feature/macos-recursive-search
-    paths:
-      - .github/workflows/apply-macos-recursive-search.yml
-      - scripts/apply_macos_recursive_search_patch.py
-
-permissions:
-  contents: write
-
-jobs:
-  apply:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout feature branch
-        uses: actions/checkout@v4
-        with:
-          ref: feature/macos-recursive-search
-          fetch-depth: 0
-      - name: Apply guarded source patch
-        run: python3 scripts/apply_macos_recursive_search_patch.py
-      - name: Remove one-shot patch machinery
-        run: |
-          rm scripts/apply_macos_recursive_search_patch.py
-          rm .github/workflows/apply-macos-recursive-search.yml
-      - name: Commit production changes only
-        run: |
-          git config user.name "Brendigo"
-          git config user.email "info@brendigo.com"
-          git add -A
-          git diff --cached --check
-          git diff --cached --quiet && { echo "Patch produced no changes" >&2; exit 1; }
-          git commit -m "feat: add native bounded macOS recursive search"
-          git push origin HEAD:feature/macos-recursive-search
